+++ src/main/java/net/decodex/loghub/backend/domain/dto/requests/CreateLogSourceDto.java
@@ -25,7 +25,7 @@ public class CreateLogSourceDto implements Serializable {
     private String ipAddress = "unknown";
     private CpuInfo cpuInfo;
     private String macAddress = "unknown";
-    private int maxRam = 0;
+    private long maxRam = 0;
     private String architecture = "unknown";
     private String releaseId = "unknown";
 }
